fix model

+++ app/src/main/java/com/test/testhttprequests/model/WeatherMain.java
@@ -5,8 +5,8 @@ import com.google.gson.annotations.SerializedName;
 public class WeatherMain {
 
     private double temp;
-    private int pressure;
-    private int humidity;
+    private double pressure;
+    private double humidity;
 
     @SerializedName("temp_min")
     private  double tempMin;
@@ -22,19 +22,19 @@ public class WeatherMain {
         this.temp = temp;
     }
 
-    public int getPressure() {
+    public double getPressure() {
         return pressure;
     }
 
-    public void setPressure(int pressure) {
+    public void setPressure(double pressure) {
         this.pressure = pressure;
     }
 
-    public int getHumidity() {
+    public double getHumidity() {
         return humidity;
     }
 
-    public void setHumidity(int humidity) {
+    public void setHumidity(double humidity) {
         this.humidity = humidity;
     }
